Refactoring 1 Decomposition
