+++ client/src/pages/Review/Document.tsx
@@ -1,0 +1,14 @@
+
+const Document = () => {
+  return (
+    <>
+      <img
+        src="../pdf.png"
+        alt="the pdf viewer"
+        className="border-x border-y border-dashed border-gray-400 rounded-3xl w-full mt-2"
+      />
+    </>
+  );
+};
+
+export default Document;
